refactor(dependencies): remove TESTING flag and streamline dependency fetching logic feat(tests): add synthetic dependency graph generation and storybook examples
